Add files via upload
The Processor is a 32-bit RISC CPU with 32-bit Address space with RISC type instructions. Hazard detection, Forwarding (from EX, MEM and WB stage outputs) and the Stall unit for Load and Branch instructions are implemented. The Instruction Cache is Direct-Mapped with a Block Size of Four Words. The main Instruction Memory is implemented in Block RAM. The Processor supports basic Arithmetic and Logic instructions, LUI, Branch Instructions, and Load-Store instructions.